Replace rpc calls from gui with direct function calls
    Remove the unused calls to vote, addpoll and addattachment.
    Replace the rpc call to rain with a direct function call.
    Split the rain rpc function into a rpc call 'rain' and a function
    'executeRain'.
