+++ src/contract/contract.cpp
@@ -6,6 +6,8 @@
 #include "keystore.h"
 #include "beacon.h"
 
+double GetTotalBalance();
+
 std::string GetBurnAddress() { return fTestNet ? "mk1e432zWKH1MW57ragKywuXaWAtHy1AHZ" : "S67nL4vELWwdDVzjgtEP4MxryarTZ9a8GB";
                              }
 
@@ -116,3 +118,95 @@ bool SignBlockWithCPID(const std::string& sCPID, const std::string& sBlockHash, 
     return true;
 }
 
+int64_t AmountFromDouble(double dAmount)
+{
+    if (dAmount <= 0.0 || dAmount > MAX_MONEY)        throw JSONRPCError(RPC_TYPE_ERROR, "Invalid amount");
+    int64_t nAmount = roundint64(dAmount * COIN);
+    if (!MoneyRange(nAmount))         throw JSONRPCError(RPC_TYPE_ERROR, "Invalid amount");
+    return nAmount;
+}
+
+std::string executeRain(std::string sRecipients)
+{
+    CWalletTx wtx;
+    wtx.mapValue["comment"] = "Rain";
+    set<CBitcoinAddress> setAddress;
+    vector<pair<CScript, int64_t> > vecSend;
+    std::string sRainCommand = ExtractXML(sRecipients,"<RAIN>","</RAIN>");
+    std::string sRainMessage = MakeSafeMessage(ExtractXML(sRecipients,"<RAINMESSAGE>","</RAINMESSAGE>"));
+    std::string sRain = "<NARR>Project Rain: " + sRainMessage + "</NARR>";
+
+    if (!sRainCommand.empty())
+        sRecipients = sRainCommand;
+
+    wtx.hashBoinc = sRain;
+    int64_t totalAmount = 0;
+    double dTotalToSend = 0;
+    std::vector<std::string> vRecipients = split(sRecipients.c_str(),"<ROW>");
+    LogPrintf("Creating Rain transaction with %" PRId64 " recipients. ", vRecipients.size());
+
+    for (unsigned int i = 0; i < vRecipients.size(); i++)
+    {
+        std::string sRow = vRecipients[i];
+        std::vector<std::string> vReward = split(sRow.c_str(),"<COL>");
+
+        if (vReward.size() > 1)
+        {
+            std::string sAddress = vReward[0];
+            std::string sAmount = vReward[1];
+
+            if (sAddress.length() > 10 && sAmount.length() > 0)
+            {
+                double dAmount = RoundFromString(sAmount,4);
+                if (dAmount > 0)
+                {
+                    CBitcoinAddress address(sAddress);
+                    if (!address.IsValid())
+                        throw JSONRPCError(RPC_INVALID_ADDRESS_OR_KEY, string("Invalid Gridcoin address: ")+sAddress);
+
+                    if (setAddress.count(address))
+                        throw JSONRPCError(RPC_INVALID_PARAMETER, string("Invalid parameter, duplicated address: ")+sAddress);
+
+                    setAddress.insert(address);
+                    dTotalToSend += dAmount;
+                    int64_t nAmount = AmountFromDouble(dAmount);
+                    CScript scriptPubKey;
+                    scriptPubKey.SetDestination(address.Get());
+                    totalAmount += nAmount;
+                    vecSend.push_back(make_pair(scriptPubKey, nAmount));
+                }
+            }
+        }
+    }
+
+    EnsureWalletIsUnlocked();
+    // Check funds
+    double dBalance = GetTotalBalance();
+
+    if (dTotalToSend > dBalance)
+        throw JSONRPCError(RPC_WALLET_INSUFFICIENT_FUNDS, "Account has insufficient funds");
+    // Send
+    CReserveKey keyChange(pwalletMain);
+    int64_t nFeeRequired = 0;
+    bool fCreated = pwalletMain->CreateTransaction(vecSend, wtx, keyChange, nFeeRequired);
+    LogPrintf("Transaction Created.");
+
+    if (!fCreated)
+    {
+        if (totalAmount + nFeeRequired > pwalletMain->GetBalance())
+            throw JSONRPCError(RPC_WALLET_INSUFFICIENT_FUNDS, "Insufficient funds");
+        throw JSONRPCError(RPC_WALLET_ERROR, "Transaction creation failed");
+    }
+
+    LogPrintf("Committing.");
+    // Rain the recipients
+    if (!pwalletMain->CommitTransaction(wtx, keyChange))
+    {
+        LogPrintf("Commit failed.");
+
+        throw JSONRPCError(RPC_WALLET_ERROR, "Transaction commit failed");
+    }
+    std::string sNarr = "Rain successful:  Sent " + wtx.GetHash().GetHex() + ".";
+    LogPrintf("Success %s",sNarr.c_str());
+    return sNarr;
+}
