+++ src/contract/contract.h
@@ -8,4 +8,4 @@ std::string SignMessage(std::string sMsg, std::string sPrivateKey);
 bool VerifyCPIDSignature(std::string sCPID, std::string sBlockHash, std::string sSignature);
 bool SignBlockWithCPID(const std::string& sCPID, const std::string& sBlockHash, std::string& sSignature, std::string& sError, bool bAdvertising=false);
 bool CheckMessageSignature(std::string sAction,std::string messagetype, std::string sMsg, std::string sSig, std::string opt_pubkey);
-
+std::string executeRain(std::string sRecipients);
